access modifier

+++ multi.java
@@ -1,16 +1,17 @@
 public class multi {
 
 
-    public static float product(int a, int b) {
-        float multiply =a*b ;
+    public static int product(int a, int b) {
+        int multiply =a*b ;
         return multiply;
     }
 
     public static void main(String[] args) {
 //        int a = add(23,1);
 //        System.out.println(a);
-        float of=product(25,15);
+        int of=product(25,15);
         System.out.println(of);
+
